+++ himotoki/cli.py
@@ -25,7 +25,7 @@ from himotoki.characters import romanize_word
 # Constants
 # =============================================================================
 
-VERSION = '0.2.0'
+VERSION = '0.3.0'
 CONJUGATION_ROOT = 'root'  # Marker for unconjugated (dictionary) form